requirements.txt
streamlit
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.3
langchain-text-splitters==0.0.1
faiss-cpu
pypdf
python-dotenv
sentence-transformers
streamlit_option_menu